webapi: Added to parameter to for character statistics
